requirements-dev.txt
Faker==9.6.0
flake8==3.8.4
mccabe==0.6.1
model-mommy==2.0.0
pycodestyle==2.6.0
pyflakes==2.2.0

# Deps do pytest
attrs==20.3.0
py==1.10.0
pluggy==0.13.1
pytest-mock==3.3.1

#Coverage
coverage==5.3
codecov==2.1.11
pytest-cov==2.10.1
pytest-django==4.1.0


-r requirements.txt
